Refactor software orchestrator module into orchestrators directory
