+++ concrete/orchestrators/base.py
@@ -1,0 +1,2 @@
+class Orchestrator:
+    pass
